% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sort_spc_summary_table}
\alias{sort_spc_summary_table}
\title{Sorting SPC summary tab.}
\usage{
sort_spc_summary_table(data, sort_by)
}
\arguments{
\item{data}{The data is the SPC summary table dataframe once icons have been determined}

\item{sort_by}{The desired sorting mechanism. Can be any of the following: "indicator", "assurance concern", "assurance improve", "variation concern", "variation improve", "improve both by variation", "improve both by assurance", "concern both by assurance", "concern both by variation"}
}
\description{
The function used to sort the SPC summary table by desired columns and desired order. Function used within the processing of the SPC summary table as columns cannot be custom sorted once the flextable or reactable is prodcued
}
\examples{

#Not to be used independently

}
