% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_chart.R
\name{spc_chart}
\alias{spc_chart}
\title{SPC Charting functionality.}
\usage{
spc_chart(
  .data,
  .base_date_range = NULL,
  .package = NULL,
  .plot_title = NULL,
  .yrange = NULL,
  .line_breaks = F,
  .chart_theme = NULL,
  .time_unit = NULL
)
}
\arguments{
\item{.data}{Data cleaned and processed within spc_output(), or data returned from spc_ouput(output = "chart").}

\item{.base_date_range}{Base date range entered into spc_output(base_date_range) or can be entered into spc_chart() the same way}

\item{.package}{Package as entered into spc_output(package) or can be entered into spc_chart() the same way ("echarts"/"echarts4r", "plotly" or "ggplot")}

\item{.plot_title}{Provides the plot with a title if the output is chart}

\item{.yrange}{Provides a axis range if the output is chart. Argument should be c(min,max). A count of 5 is +/- from values}

\item{.line_breaks}{Determines whether mean and process limit lines break after a rebase. Defaults to F description}

\item{.chart_theme}{Takes a list of arguments that would appear in theme argument of}

\item{.time_unit}{If time unit is set to "day", "week" or "month", reflecting daily, weekly, monthly date periods. if set to "quarter", x-axis becomes factor and converts the axis to quarterly labels.}
}
\description{
The function used to output data as an SPC chart. Can be ran independently from spc_output()
}
\details{
Use this function on a spc processed data frame. Can only provide graph for one metric upon each call (doesn't support faceted charts as of yet)

append \%>\% spc_add_icons to end of chart object to add icons.
}
\examples{
#An SPC Chart

library(dplyr)

start_date <- as.Date("2025-01-01")
end_date <- as.Date("2025-01-20")
date_sequence <- seq.Date(from = start_date, to = end_date, by = "day")


indicator_data_1 <- 
  data.frame(Date = date_sequence,
             kpi = "Indicator 1",
             indicator_value = c(45,48,44,43,45,
                                 65,45,46,46,44,
                                 43,42,41,40,39,
                                 46,47,56,52,50),
             target = 60)

indicator_data_2 <- 
  data.frame(Date = date_sequence,
             kpi = "Indicator 2",
             indicator_value = c(45,48,44,43,45,
                                 45,45,47,46,44,
                                 43,44,43,28,44,
                                 45,47,45,43,46),
             target = 45)


all_indicator_data <- bind_rows(indicator_data_1, indicator_data_2) \%>\% 
  dplyr::mutate(polarity = "up",
                greater_than_hundred = FALSE,
                less_than_zero = FALSE,
                unit = "count")

spc_data <- HertsSPC::spc_output(
  data = all_indicator_data,
  time_field = "Date",
  indicator = "kpi",
  value = "indicator_value",
  output = "data",
  target = "target"
)



spc_chart(.data = filter(spc_data, indicator == "Indicator 1"),
          .package = "ggplot",
          .plot_title = "SPC for Indicator 1 (ggplot2)",
          .base_date_range = NULL,
          .chart_theme = spc_chart_options(x_title_size = 20,
                                  x_label = "Date",
                                  y_label = "Measure",
                                  x_label_format = "\%d \%b",
                                  x_breaks = "1 day"))
          

spc_chart(.data = filter(spc_data, indicator == "Indicator 2"),
          .plot_title = "SPC for Indicator 2 (Plotly)",
          .base_date_range = NULL,
          .package = "plotly",
          .yrange = c(0, 100))



}
