% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_chart.R
\name{spc_chart_options}
\alias{spc_chart_options}
\title{SPC Chart options}
\usage{
spc_chart_options(
  x_title_size = NULL,
  x_text_size = NULL,
  x_label_format = "\%b-\%Y",
  y_title_size = NULL,
  y_text_size = NULL,
  y_label = NULL,
  x_label = NULL,
  title_size = NULL,
  x_breaks = "1 month"
)
}
\arguments{
\item{x_title_size}{Size of x axis title}

\item{x_text_size}{Size of x axis tick labels}

\item{x_label_format}{e.g. "\%b \%d"}

\item{y_title_size}{Size of y axis title}

\item{y_text_size}{Size of y axis tick labels}

\item{y_label}{y axis title}

\item{x_label}{x axis title}

\item{title_size}{Size of title}

\item{x_breaks}{Breaks in x axis. Defaults to 3 months}
}
\description{
Chart options for ggplot. Pipe spc_add_icons() onto end of function to apply icons (sizes can be ammended to suit the plot if needed, see ?spc_add_icons.
}
\examples{

library(dplyr)

start_date <- as.Date("2025-01-01")
end_date <- as.Date("2025-01-20")
date_sequence <- seq.Date(from = start_date, to = end_date, by = "day")

indicator_data <-
  data.frame(Date = date_sequence,
             kpi = "Indicator 1",
             indicator_value = c(45,48,44,43,45,
                                 65,45,46,46,44,
                                 43,42,41,40,39,
                                 46,47,56,52,50),
             target = 60)

spc_output(
  data = indicator_data,
  time_field = "Date",
  indicator = "kpi",
  value = "indicator_value",
  output = "chart",
  package = "ggplot",
  chart_theme = spc_chart_options(x_title_size = 20,
                                  x_label = "Date",
                                  y_label = "Measure",
                                  x_label_format = "\%d \%b",
                                  x_breaks = "1 day"))

}
