% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_icons.R
\name{spc_icons}
\alias{spc_icons}
\title{SPC icons functionality.}
\usage{
spc_icons(.data, .polarity)
}
\arguments{
\item{.data}{Data cleaned and processed within spc_output(), or data returned from spc_ouput(output = "chart").}

\item{.polarity}{The polarity of the indicator in question, obtained from column in dataframe. "up" or "down"}
}
\description{
Icons generation used for SPC charts. Need www/ folder in project directory (found in Reference and Guidance in Data Science Secondment). Cannot be run independently from spc_ouput()/spc_chart().
Returns a list of icons, element 1 being Assurance and element 2 being Variation.
}
\examples{
# NOT TO BE CALLED INDEPENDENTLY
}
