% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{spc_img_uri}
\alias{spc_img_uri}
\title{Attaining SPC icons in summary reactable table.}
\usage{
spc_img_uri(icon_string)
}
\arguments{
\item{icon_string}{the icons locations}
}
\description{
The function used to read in SPC icons for reactable summary table
}
\examples{

#Not to be used independently

}
