% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_narrative.R
\name{spc_narrative}
\alias{spc_narrative}
\title{SPC Narrative functionality.}
\usage{
spc_narrative(.data, .mode, .time_unit = "month")
}
\arguments{
\item{.data}{Data cleaned and processed within spc_output(), or data returned from spc_ouput(output = "data"). Can only be run with one indicator unless function is in loop.}

\item{.mode}{"static" or "interactive".}

\item{.time_unit}{Defaults to "month". Can be "day", "month", "quarter" or "week".}
}
\description{
The function used to output data as an SPC chart. Can be ran independently from spc_output()
}
\examples{

library(dplyr)

start_date <- as.Date("2025-01-01")
end_date <- as.Date("2025-01-20")
date_sequence <- seq.Date(from = start_date, to = end_date, by = "day")


indicator_data_1 <- 
  data.frame(Date = date_sequence,
             kpi = "Indicator 1",
             indicator_value = c(45,48,44,43,45,
                                 65,45,46,46,44,
                                 43,42,41,40,39,
                                 46,47,56,52,50),
             target = 60)

indicator_data_2 <- 
  data.frame(Date = date_sequence,
             kpi = "Indicator 2",
             indicator_value = c(45,48,44,43,45,
                                 45,45,47,46,44,
                                 43,44,43,28,44,
                                 45,47,45,43,60),
             target = 45)


all_indicator_data <- bind_rows(indicator_data_1, indicator_data_2) \%>\% 
  dplyr::mutate(polarity = "up",
                greater_than_hundred = FALSE,
                less_than_zero = FALSE,
                unit = "count")
                
spc_data <- HertsSPC::spc_output(
  data = all_indicator_data,
  time_field = "Date",
  indicator = "kpi",
  value = "indicator_value",
  output = "data",
  target = "target"
)              

spc_narrative(.data = filter(spc_data, indicator == "Indicator 1"),
              .mode = "static",
              .time_unit = "day")

spc_narrative(.data = filter(spc_data, indicator == "Indicator 2"),
              .mode = "interactive",
              .time_unit = "day")
              
HertsSPC::spc_output(
  data = filter(all_indicator_data, kpi == "Indicator 1"),
  time_field = "Date",
  indicator = "kpi",
  value = "indicator_value",
  output = "narrative",
  mode = "interactive",
  target = "target"
)                     
              
}
