% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_output.R
\name{spc_output}
\alias{spc_output}
\title{Main function to return desired statistical process output}
\usage{
spc_output(
  data,
  indicator,
  time_field,
  value,
  exclude_outliers = FALSE,
  baseline_point_number = NULL,
  base_date_range = NULL,
  rebase_dates = NULL,
  rebase_data_frame = NULL,
  target = NULL,
  mode = "static",
  output = "chart",
  summary_output = "table",
  package = NULL,
  plot_title = NULL,
  yrange = NULL,
  group_average = FALSE,
  line_breaks = FALSE,
  chart_theme = NULL,
  nad = FALSE,
  sort_by = "concern both by assurance",
  time_unit = "month"
)
}
\arguments{
\item{data}{A df with the following columns: 1. A indicator column, e.g. KPI (used to group by if processing more than one indicator); 2. A time field (dates); 3. A value field reflecting performance; 4. A polarity field indicting whether an upward trend indicates improvement ("up", "down", "neutral"); 5. A field called greater_than_hundred indicating whether the value can be legitimately over 100; 6. A field called less_than_zero indicating whether the value can be below zero (T / F); 7. A unit column indicating type of data ("percent" or "count")}

\item{indicator}{The column in data reflecting the group by column, as a character string. If no grouping column exists, set as NULL.}

\item{time_field}{The column in data reflecting the time component, as a character string}

\item{value}{The column in data reflecting the value component to be measured, as a character string}

\item{exclude_outliers}{Set to T if outliers are desired to be excluded from process calculations
Reflects periods of rebasing. Can be rebased multiple times. Processing function will fill dates of rebasing downwards. E.g. above example, there will be three groups, start:2021-09-01, 2021-10-01:2022-06-01, 2022-07-01:end}

\item{baseline_point_number}{Defaults to NULL Set to a number if you want your SPC to be calculated using only the first n points of your data (and each distinct rebased period). This sets the baseline to n points across each rebased group. Conflicts with base_date_range, however can be used in conjunction with rebase_dates or rebase_data_frame}

\item{base_date_range}{A base date range, structure like c("2021-02-01", "2021-09-01") i.e. c(start, end)}

\item{rebase_dates}{Manually decided rebasing of dates. Structured like c("2021-09-01", "2022-06-01").}

\item{rebase_data_frame}{Defaults to NULL. Assign dataframe of interest if you have a dataframe of rebase dates by indicator. Dataframe will join by indicator, meaning unique rebase dates can be applied on mass to multiple indicators during processing. If indicator does not exist in rebasing df, then no rebasing will occur to particular indicator.}

\item{target}{Defaults to NULL. Target can either be inputted as a column character string or a known integer, e.g. 80}

\item{mode}{"interactive" or "static", depending on output desired. Applied to narratives and summaries}

\item{output}{"data", "chart", "summary", "narrative" or "status"}

\item{summary_output}{The desired output type of summary, depending on intentions. Either "table", which produces the final table summary (flextable or reactable depending on mode), or "dataframe", which returns the summary as is before the final table output. Allows for further editing.}

\item{package}{In respect to interactive charts output, "ggplot2", "echarts"/"echarts4r" or "plotly"}

\item{plot_title}{Provides the plot with a title if the output is chart}

\item{yrange}{Provides a axis range if the output is chart. Argument should be c(min,max). A count of 5 is +/- from values}

\item{group_average}{Compare a metric value against the process of the average of a group of metrics. Defaults to F.}

\item{line_breaks}{Defaults to F. Determines whether lines will have breaks between rebasing (LCL, UCL and Mean)}

\item{chart_theme}{Theme for chart. Uses spc_chart_options function (see ?HccSPC::spc_chart_options()).}

\item{nad}{Numerator and denominator. Set to TRUE if you have a Numerator and Denominator column and want to present them figures in the table}

\item{sort_by}{Determines sorting of table. Can be sorted by assurance, variation, by both, or by alphabetical order of indicator. Defaults to "concern both by assurance". See ?HertsSPC::sort_spc_summary_table() for all available options. To sort by indicator alphabetically, assign "indicator" description}

\item{time_unit}{Defaults to "month". Can set as "day", "month", "quarter" or "week". Only relevant when using summary and narrative outputs. Chart only impacted if "quarter" is selected.}
}
\description{
A general function used to return a desired statistical process output without needing to call individual output functions.

Dataframes going into this SPC function need te following columns (and if they do not exist, they will be defaulted within the function):
\itemize{
\item unit (either "count" of "percent")
\item greater_than_hundred (TRUE or FALSE). This implies whether the KPI/metric can theoretically exceed 100 (as a count or as a percentage) for graphical purposes
\item less_than_zero (TRUE or FALSE). This implies whether the KPI/metric can theoretically go below 0 (as a count or as a percentage) for graphical purposes
\item polarity ("down" or "up"). Determines whether an improvement is up or down
\item A time field
\item A value field
\item An indicator field
\item Target (optional)
}

Having columns for these indicators allows producing SPC charts/tables on mass with indicators of varying properties
}
\examples{

library(dplyr)

start_date <- as.Date("2025-01-01")
end_date <- as.Date("2025-01-20")
date_sequence <- seq.Date(from = start_date, to = end_date, by = "day")


indicator_data_1 <- 
  data.frame(Date = date_sequence,
             kpi = "Indicator 1",
             indicator_value = c(45,48,44,43,45,
                                 65,45,46,46,44,
                                 43,42,41,40,39,
                                 46,47,56,52,50),
             target = 60)


#Retreive SPC data

spc_output(
       data = indicator_data_1,
       time_field = "Date",
       indicator = "kpi",
       value = "indicator_value",
       output = "data")

# Retrieve static SPC plot

indicator_data_3 <- 
     data.frame(
        Date = seq.Date(from = start_date, 
                        to = as.Date("2025-01-30"), 
                        by = "day"),
        kpi = "Indicator 3",
        indicator_value = c(45,48,44,43,45,
                            45,45,47,46,44,
                                 43,44,43,28,44,
                                 85,88,88,83,85,
                                 85,85,87,86,88,
                                 83,88,83,88,88),
             target = 45)

spc_output(
data = indicator_data_3,
rebase_dates = "2021-01-15", #With a rebase date,
time_field = "Date",
indicator = "kpi",
value = "indicator_value",
output = "chart",
package = "ggplot") \%>\%
spc_add_icons() # with icons


spc_output(
data = indicator_data_3,
rebase_dates = c("2021-01-10", "2021-01-20"),
time_field = "Date",
indicator = "kpi",
value = "indicator_value",
output = "chart",
package = "ggplot",
target = 30)


spc_output(
       data = indicator_data_1,
       time_field = "Date",
       indicator = "kpi",
       value = "indicator_value",
       output = "chart",
       package = "ggplot",
       target = 30,
       plot_title = "Performance over time",
       time_unit = "day",
       chart_theme = spc_chart_options(x_label = "Day",
                                       y_label = "Count",
                                       x_label_format = "\%b-\%d",
                                       title_size = 25,
                                       x_breaks = "1 day"
                                       ))

# Retrieve SPC in plotly

spc_output(
       data = indicator_data_1,
       time_field = "Date",
       indicator = "kpi",
       value = "indicator_value",
       output = "chart",
       package = "plotly")

# Retrieve SPC in echarts

spc_output(
       data = indicator_data_1,
       time_field = "Date",
       indicator = "kpi",
       value = "indicator_value",
       output = "chart",
       package = "echarts") \%>\%
       spc_add_icons()


# Retrieve a narrative

spc_output(
       data = indicator_data_1,
       time_field = "Date",
       indicator = "kpi",
       value = "indicator_value",
       output = "narrative")

spc_output(
       data = indicator_data_1,
       time_field = "Date",
       indicator = "kpi",
       value = "indicator_value",
       output = "narrative",
       mode = "interactive")


# Retrieve summary table

spc_output(
       data = indicator_data_1,
       time_field = "Date",
       indicator = "kpi",
       value = "indicator_value",
       output = "summary",
       mode = "static")

spc_output(
       data = dplyr::bind_rows(indicator_data_1,  indicator_data_3),
       time_field = "Date",
       indicator = "kpi",
       value = "indicator_value",
       output = "summary",
       mode = "interactive",
       nad = FALSE)


}
