% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_crosstalk.R
\name{spc_plotly_crosstalk}
\alias{spc_plotly_crosstalk}
\title{SPC Charting functionality.}
\usage{
spc_plotly_crosstalk(data)
}
\arguments{
\item{data}{Data is the output of spc_plotly_data() (a SharedData object)}
}
\description{
The function used to output data as a plotly chart linked to a crosstalk filter.
}
\examples{

library(dplyr)

tooth_data <- force(ToothGrowth) \%>\%
  filter(supp == "VC") \%>\% slice(-15:-27) \%>\%
  mutate(Date = seq.Date(as.Date("2021-01-01"), as.Date("2021-01-17"), by = "days"),
         supp = "Indicator 1",
         polarity = "up",
         greater_than_hundred = FALSE,
         less_than_zero = FALSE,
         unit = "count")

tooth_data2 <- force(ToothGrowth) \%>\%
  filter(supp == "VC") \%>\% slice(-15:-27) \%>\%
  mutate(Date = seq.Date(as.Date("2021-01-01"), as.Date("2021-01-17"), by = "days"),
         supp = "Indicator 2",
         polarity = "up",
         greater_than_hundred = FALSE,
         less_than_zero = FALSE,
         unit = "count") \%>\%
  bind_rows(tooth_data)


# Retreive SPC data

spc_data_for_crosstalk <- spc_plotly_data(
       data = tooth_data2,
       time_field = "Date",
       indicator = "supp",
       value = "len")

spc_plotly_crosstalk(spc_data_for_crosstalk)

}
