% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_crosstalk.R
\name{spc_plotly_data}
\alias{spc_plotly_data}
\title{SPC Charting functionality.}
\usage{
spc_plotly_data(
  data,
  indicator,
  time_field,
  value,
  exclude_outliers = F,
  base_date_range = NULL,
  rebase_dates = NULL,
  rebase_data_frame = NULL,
  target = NULL,
  group_average = F
)
}
\arguments{
\item{data}{A df with the following columns: 1. A indicator column, e.g. KPI (used to group by if processing more than one indicator); 2. A time field (dates); 3. A value field reflecting performance; 4. A polarity field indicting whether an upward trend indicates improvement ("up", "down", "neutral"); 5. A field called greater_than_hundred indicating whether the value can be legitimately over 100; 6. A field called less_than_zero indicating whether the value can be below zero (T / F); 7. A unit column indicating type of data ("percent" or "count")}

\item{indicator}{The column in data reflecting the group by column, as a character string. If no grouping column exists, set as NULL.}

\item{time_field}{The column in data reflecting the time component, as a character string}

\item{value}{The column in data reflecting the value component to be measured, as a character string}

\item{exclude_outliers}{Set to T if outliers are desired to be excluded from process calculations}

\item{base_date_range}{A base date range, structure like c("2021-02-01", "2021-09-01") i.e. c(start, end)}

\item{rebase_dates}{Manually decided rebasing of dates. Structured like c("2021-09-01", "2022-06-01").}

\item{rebase_data_frame}{Defaults to NULL. Assign dataframe of interest if you have a dataframe of rebase dates by indicator. Dataframe will join by indicator, meaning unique rebase dates can be applied on mass to multiple indicators during processing. If indicator does not exist in rebasing df, then no rebasing will occur to particular indicator.}

\item{target}{Defaults to NULL. Target can either be inputted as a column character string or a known integer, e.g. 80}

\item{group_average}{Compare a metric value against the process of the average of a group of metrics. Defaults to F.}
}
\description{
The function used to create an object a SPC datafame to be used as a crosstalk shared object. Creates a dataframe to be used a s a SharedData. Includes less arguements as normal SPC output/chart functionality as crosstalk is relatively restrcitive (e.g. no icons)
}
\examples{


library(dplyr)

tooth_data <- force(ToothGrowth) \%>\%
  filter(supp == "VC") \%>\% slice(-15:-27) \%>\%
  mutate(Date = seq.Date(as.Date("2021-01-01"), as.Date("2021-01-17"), by = "days"),
         supp = "Indicator 1",
         polarity = "up",
         greater_than_hundred = FALSE,
         less_than_zero = FALSE,
         unit = "count")

tooth_data2 <- force(ToothGrowth) \%>\%
  filter(supp == "VC") \%>\% slice(-15:-27) \%>\%
  mutate(Date = seq.Date(as.Date("2021-01-01"), as.Date("2021-01-17"), by = "days"),
         supp = "Indicator 2",
         polarity = "up",
         greater_than_hundred = FALSE,
         less_than_zero = FALSE,
         unit = "count") \%>\%
  bind_rows(tooth_data)


# Retreive SPC data

spc_data_for_crosstalk <- spc_plotly_data(
       data = tooth_data2,
       time_field = "Date",
       indicator = "supp",
       value = "len")

}
