% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_processing.R
\name{spc_processing}
\alias{spc_processing}
\title{SPC Processing}
\usage{
spc_processing(
  data,
  time_field,
  value,
  base_date_range = NULL,
  rebase_dates = NULL,
  rebase_data_frame = NULL,
  group_average = F,
  exclude_outliers = F,
  baseline_point_number = NULL,
  target = NULL
)
}
\arguments{
\item{data}{Data cleaned within the first half of spc_output(). This data is a df grouped by indicator}

\item{time_field}{The column reflecting the time field}

\item{value}{The column reflecting the value field}

\item{base_date_range}{Base date range entered into spc_output(base_date_range)}

\item{rebase_dates}{Rebase dates entered into spc_output(rebase_dates)}

\item{rebase_data_frame}{Whether rebase dates are coming from a dataframe}

\item{group_average}{Compare a metric value against the process of the average of a group of metrics. Defaults to F. Provided in spc_output().}

\item{exclude_outliers}{Set to T if outliers are desired to be excluded from process calculations. Provided in spc_output().}

\item{baseline_point_number}{baseline_point_number entered into spc_output()}

\item{target}{The column reflecting the target column}
}
\description{
The function used to process the data inputted into spc_output(). The function does not work independently
}
\examples{

#NOT TO BE CALLED INDEPENDENTLY

}
