% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{spc_reactable_tippy}
\alias{spc_reactable_tippy}
\title{Creating hover-over for reactable summary icons. Returns the variation/assurance for a given indicator when hovering cursor over icons}
\usage{
spc_reactable_tippy(text, tooltip, placement)
}
\arguments{
\item{text}{value in the reactable, in this case the icons}

\item{tooltip}{Text to be displayed in tippy (gathered from data in SPC summary function)}

\item{placement}{placement of hover-over}
}
\description{
The function used to read in SPC icons for reactable summary table
}
\examples{

#Not to be used independently

}
