% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_status.R
\name{spc_status}
\alias{spc_status}
\title{SPC status table functionality.}
\usage{
spc_status(.data, .indicator, .time_field, .value)
}
\arguments{
\item{.data}{Data cleaned and processed within spc_output(), or data returned from spc_ouput(output = "data")}

\item{.indicator}{column reflecting the indicator, which is grouped. Needed for column header}

\item{.time_field}{column reflecting the time. Needed for column header}

\item{.value}{colunm reflecting value to be reported. Needed for column header}
}
\description{
The function used to output data as a metric level status dataframe. Includes both assurance and variation
}
\examples{

library(dplyr)

start_date <- as.Date("2025-01-01")
end_date <- as.Date("2025-01-20")
date_sequence <- seq.Date(from = start_date, to = end_date, by = "day")


indicator_data_1 <- 
  data.frame(Date = date_sequence,
             kpi = "Indicator 1",
             indicator_value = c(45,48,44,43,45,
                                 65,45,46,46,44,
                                 43,42,41,40,39,
                                 46,47,56,52,50),
             target = 60)

indicator_data_2 <- 
  data.frame(Date = date_sequence,
             kpi = "Indicator 2",
             indicator_value = c(45,48,44,43,45,
                                 45,45,47,46,44,
                                 43,44,43,28,44,
                                 45,47,45,43,46),
             target = 45)


all_indicator_data <- bind_rows(indicator_data_1, indicator_data_2) \%>\% 
  dplyr::mutate(polarity = "up",
                greater_than_hundred = FALSE,
                less_than_zero = FALSE,
                unit = "count")

spc_data <- HertsSPC::spc_output(
  data = all_indicator_data,
  time_field = "Date",
  indicator = "kpi",
  value = "indicator_value",
  output = "data",
  target = "target"
)


spc_data <- spc_output(data = all_indicator_data,
                       time_field = "Date", 
                        indicator = "kpi",
                        value = "indicator_value",
                        output = "data",
                        target = "target"
                        )

spc_status(.data = spc_data,
                  .indicator = "KPI",
                  .value = "Quantity",
                  .time_field = "Day")

spc_output(data = all_indicator_data,
                       time_field = "Date",
                        indicator = "kpi",
                        value = "indicator_value",
                        output = "status",
                        target = "target"
                        )


}
