% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_summary_table.R
\name{spc_summary_table}
\alias{spc_summary_table}
\title{SPC Summary Table functionality.}
\usage{
spc_summary_table(
  .data,
  .mode = "static",
  .indicator,
  .time_field,
  .value,
  .nad = FALSE,
  .time_unit = "month",
  .summary_output = "table",
  .sort_by = "concern both by assurance"
)
}
\arguments{
\item{.data}{Data cleaned and processed within spc_output(), or data returned from spc_ouput(output = "data").}

\item{.mode}{mode entered into spc_output(mode) or can be entered into spc_summary_table() the same way ("static" or "interactive")}

\item{.indicator}{column reflecting the indicator, which is grouped. Needed for column header}

\item{.time_field}{column reflecting the time. Needed for column header}

\item{.value}{column reflecting value to be reported. Needed for column header}

\item{.nad}{Determines whether the numerator and denominator appear in the final table (alongside the value). Only set to T if a numerator and denominator column is present in your input dataframe. Defaults to F.}

\item{.time_unit}{Can be set to "month", "day", "quarter" or "week". Determines the date column format in the summary table.}

\item{.summary_output}{summary_output The desired output type of summary, depending on intentions. Either "table", which produces the final table summary (flextable or reactable depending on mode), or "dataframe", which returns the summary as is before the final table output. Allows for further editing.}

\item{.sort_by}{Determines sorting of table. Can be sorted by assurance, variation, by both, or by alphabetical order of indicator. Defaults to "concern both by assurance". See ?HertsSPC::sort_spc_summary_table() for all available options. To sort by indicator alphabetically, assign "indicator"}
}
\description{
The function used to output data as an SPC summary table. Can be ran independently from spc_output()
}
\examples{

library(dplyr)

start_date <- as.Date("2025-01-01")
end_date <- as.Date("2025-01-20")
date_sequence <- seq.Date(from = start_date, to = end_date, by = "day")


indicator_data_1 <- 
  data.frame(Date = date_sequence,
             kpi = "Indicator 1",
             Numerator = c(45,48,44,43,45,
                                 65,45,46,46,44,
                                 43,42,41,40,39,
                                 46,47,56,52,50),
             Denominator = c(95,99,94,101,96,
                                 95,96,94,97,100,
                                 88,95,93,96,92,
                                 95,95,96,97,96),                   
             target = 60,
             unit = "percent",
             polarity = "up",
             greater_than_hundred = FALSE,
             less_than_zero = FALSE) \%>\% 
 dplyr::mutate( indicator_value = round((Numerator/Denominator)*100, 2))


indicator_data_2 <- 
  data.frame(Date = date_sequence,
             kpi = "Indicator 2",
             indicator_value = c(45,48,44,43,45,
                                 45,45,47,46,44,
                                 43,44,43,28,44,
                                 45,47,45,43,46),
             target = 45,
             greater_than_hundred = TRUE,
             unit = "count",
             polarity = "up",
             less_than_zero = FALSE)


all_indicator_data <- bind_rows(indicator_data_1, indicator_data_2)

spc_data <- HertsSPC::spc_output(
  data = all_indicator_data,
  time_field = "Date",
  indicator = "kpi",
  value = "indicator_value",
  output = "data",
  target = "target"
)


spc_data <- spc_output(data = all_indicator_data,
                       time_field = "Date", 
                        indicator = "kpi",
                        value = "indicator_value",
                        output = "data",
                        target = "target"
                        )

spc_summary_table(.data = spc_data,
                  .mode = "static",
                  .time_field = "Day",
                  .time_unit = "day",
                  .indicator = "Metric",
                  .value = "Value",
                  .nad = TRUE)

spc_summary_table(.data = spc_data,
                  .mode = "interactive",
                  .time_field = "Day",
                  .time_unit = "day",
                  .indicator = "Metric",
                  .value = "Value",
                  .nad = FALSE)

}
