% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_status.R
\name{spc_variation}
\alias{spc_variation}
\title{SPC variation functionality.}
\usage{
spc_variation(data, metric)
}
\arguments{
\item{data}{Dataframe with SPC processed metric.}

\item{metric}{The metric in the indicator column you want to return variation for.}
}
\description{
Return the state of variation for a selected metric that has been processed into an SPC dataframe.
}
\examples{

#NOT TO BE CALLED INDEPENDENTLY

}
