% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_icons.R
\name{spc_add_icons}
\alias{spc_add_icons}
\title{SPC icons functionality.}
\usage{
spc_add_icons(
  .spc,
  echarts_variation = c(245, 0, -999, 50, 50),
  echarts_assurance = c(200, 0, -999, 50, 50),
  ggplot_variation = c(0.32, 0.4, 0.1),
  ggplot_assurance = c(0.4, 0.4, 0.1),
  plotly_variation = c(0.8, 0.95, 0.1, 0.1),
  plotly_assurance = c(0.8, 0.95, 0.1, 0.1)
)
}
\arguments{
\item{.spc}{SPC chart ggplot (spc_chart()).}

\item{echarts_variation}{c(right, top, z, width, height)}

\item{echarts_assurance}{c(right, top, z, width, height)}

\item{ggplot_variation}{c(x,y,scale)}

\item{ggplot_assurance}{c(x,y,scale)}

\item{plotly_variation}{c(x,y,sizex,sizey)}

\item{plotly_assurance}{c(x,y,sizex,sizey)}
}
\description{
Add icons after your ggplot is created and edited.
}
\examples{
# NOT TO BE CALLED INDEPENDENTLY
}
