% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_status.R
\name{spc_assurance}
\alias{spc_assurance}
\title{SPC assruance functionality.}
\usage{
spc_assurance(data, metric)
}
\arguments{
\item{data}{Dataframe with SPC processed metric.}

\item{metric}{The metric in the indicator column you want to return assurance for}
}
\description{
Return the state of assurance for a selected metric that has been processed into an SPC dataframe.
}
\examples{

#NOT TO BE CALLED INDEPENDENTLY

}
