% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_chart.R
\name{spc_chart_options}
\alias{spc_chart_options}
\title{SPC Chart options}
\usage{
spc_chart_options(
  x_title_size = NULL,
  x_text_size = NULL,
  x_label_format = "\%b-\%Y",
  y_title_size = NULL,
  y_text_size = NULL,
  y_label = NULL,
  x_label = NULL,
  title_size = NULL,
  x_breaks = "1 month"
)
}
\arguments{
\item{x_title_size}{Size of x axis title}

\item{x_text_size}{Size of x axis tick labels}

\item{x_label_format}{e.g. "\%b \%d"}

\item{y_title_size}{Size of y axis title}

\item{y_text_size}{Size of y axis tick labels}

\item{y_label}{y axis title}

\item{x_label}{x axis title}

\item{title_size}{Size of title}

\item{x_breaks}{Breaks in x axis. Defaults to 3 months}
}
\description{
Chart options for ggplot. Pipe spc_add_icons() onto end of function to apply icons (sizes can be ammended to suit the plot if needed, see ?spc_add_icons.
}
\examples{

library(dplyr)

tooth_data <- force(ToothGrowth) \%>\%
  filter(supp == "VC") \%>\% slice(-15:-27) \%>\%
  mutate(Date = seq.Date(as.Date("2021-01-01"), as.Date("2021-01-17"), by = "days"),
         supp = "Indicator 1",
         polarity = "up",
         greater_than_hundred = FALSE,
         less_than_zero = FALSE,
         unit = "count")

spc_output(
       data = tooth_data,
       time_field = "Date",
       indicator = "supp",
       value = "len",
       output = "chart",
       package = "ggplot",
       chart_theme = spc_chart_options(x_title_size = 20,
                                        x_label = "Date",
                                        y_label = "Measure",
                                        x_label_format = "\%d \%b"))

}
