% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_narrative.R
\name{spc_narrative}
\alias{spc_narrative}
\title{SPC Narrative functionality.}
\usage{
spc_narrative(.data, .mode, .time_unit = "month")
}
\arguments{
\item{.data}{Data cleaned and processed within spc_output(), or data returned from spc_ouput(output = "data"). Can only be run with one indicator unless function is in loop.}

\item{.mode}{"static" or "interactive".}

\item{.time_unit}{Defaults to "month". Can be "day", "month", "quarter" or "week".}
}
\description{
The function used to output data as an SPC chart. Can be ran independently from spc_output()
}
\examples{

library(dplyr)

tooth_data <- force(ToothGrowth) \%>\%
  filter(supp == "VC") \%>\% slice(-15:-27) \%>\%
  mutate(Date = seq.Date(as.Date("2021-01-01"), as.Date("2021-01-17"), by = "days"),
         supp = "Indicator 1",
         polarity = "up",
         greater_than_hundred = FALSE,
         less_than_zero = FALSE,
         unit = "count")


spc_data <- spc_output(data = tooth_data,
                       time_field = "Date",
                       indicator = "supp",
                       value = "len",
                       output = "data")

spc_narrative(.data = spc_data,
              .mode = "static")

spc_narrative(.data = spc_data,
              .mode = "interactive")
}
