% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_status.R
\name{spc_status}
\alias{spc_status}
\title{SPC status table functionality.}
\usage{
spc_status(.data, .indicator, .time_field, .value)
}
\arguments{
\item{.data}{Data cleaned and processed within spc_output(), or data returned from spc_ouput(output = "data")}

\item{.indicator}{column reflecting the indicator, which is grouped. Needed for column header}

\item{.time_field}{column reflecting the time. Needed for column header}

\item{.value}{colunm reflecting value to be reported. Needed for column header}
}
\description{
The function used to output data as a metric level status dataframe. Includes both assurance and variation
}
\examples{

library(dplyr)

tooth_data <- force(ToothGrowth) \%>\%
  filter(supp == "VC") \%>\% slice(-15:-27) \%>\%
  mutate(Date = seq.Date(as.Date("2021-01-01"), as.Date("2021-01-17"), by = "days"),
         supp = "Indicator 1",
         polarity = "up",
         greater_than_hundred = FALSE,
         less_than_zero = FALSE,
         unit = "count")


spc_data <- spc_output(data = tooth_data,
                       time_field = "Date",
                       indicator = "supp",
                       value = "len",
                       output = "data")

spc_status(.data = spc_data,
                  .indicator = "Tooth Type",
                  .value = "Size",
                  .time_field = "Day")

}
