% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_summary_table.R
\name{spc_summary_table}
\alias{spc_summary_table}
\title{SPC Summary Table functionality.}
\usage{
spc_summary_table(
  .data,
  .mode = "static",
  .indicator,
  .time_field,
  .value,
  .nad = FALSE,
  .time_unit = "month"
)
}
\arguments{
\item{.data}{Data cleaned and processed within spc_output(), or data returned from spc_ouput(output = "data").}

\item{.mode}{mode entered into spc_output(mode) or can be entered into spc_summary_table() the same way ("static" or "interactive")}

\item{.indicator}{column reflecting the indicator, which is grouped. Needed for column header}

\item{.time_field}{column reflecting the time. Needed for column header}

\item{.value}{column reflecting value to be reported. Needed for column header}

\item{.nad}{Determines whether the numerator and denominator appear in the final table (alongside the value). Only set to T if a numerator and denominator column is present in your input dataframe. Defaults to F.}

\item{.time_unit}{Can be set to "month", "day", "quarter" or "week". Determines the date column format in the summary table.}
}
\description{
The function used to output data as an SPC summary table. Can be ran independently from spc_output()
}
\examples{

library(dplyr)

tooth_data <- force(ToothGrowth) \%>\%
  filter(supp == "VC") \%>\% slice(-15:-27) \%>\%
  mutate(Date = seq.Date(as.Date("2021-01-01"), as.Date("2021-01-17"), by = "days"),
         supp = "Indicator 1",
         polarity = "up",
         greater_than_hundred = FALSE,
         less_than_zero = FALSE,
         unit = "count")


spc_data <- spc_output(data = tooth_data,
                       time_field = "Date",
                       indicator = "supp",
                       value = "len",
                       output = "data")

spc_summary_table(.data = spc_data,
                  .mode = "static",
                  .time_field = "Day",
                  .indicator = "Tooth Type",
                  .value = "Size",
                  .nad = FALSE)

spc_summary_table(.data = spc_data,
                  .mode = "interactive",
                  .time_field = "Day",
                  .indicator = "Tooth Type",
                  .value = "Size",
                  .nad = FALSE)

}
